Refactor events
